add voice_controller in main.py
