+++ launch/robot_lab.launch
@@ -27,4 +27,7 @@
   <node name="joy" pkg="joy" type="joy_node">
     <param name="dev_name" value="Sony Playstation SixAxis/DS3" />
   </node>
+
+  <node name="speech" pkg="voice_control" type="speech.py"/>
+  <node name="controller" pkg="voice_control" type="controller.py"/>
 </launch>
